+++ src/components/app/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from "prop-types";
-import WelcomeScreen from '../welcome-screen/WelcomeScreen';
+import WelcomeScreen from '../welcome-screen/welcome-screen';
 
 const App = ({errorsCount}) => {
   return (
