+++ src/components/welcome-screen/welcome-screen.jsx
